working lazy list for showing PokemonList in MainActivity
