+++ app/src/main/AndroidManifest.xml
@@ -13,6 +13,11 @@
         android:theme="@style/Theme.PokemonDex"
         tools:targetApi="31">
         <activity
+            android:name=".PokeDetail"
+            android:exported="false"
+            android:label="@string/title_activity_poke_detail"
+            android:theme="@style/Theme.PokemonDex" />
+        <activity
             android:name=".MainActivity"
             android:exported="true"
             android:label="@string/app_name"
