+++ app/src/main/java/com/krisna/pokemondex/MainActivity.kt
@@ -3,13 +3,23 @@ package com.krisna.pokemondex
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import com.krisna.pokemondex.model.Pokemon
+import com.krisna.pokemondex.model.PokemonData
 import com.krisna.pokemondex.ui.theme.PokemonDexTheme
 
 
@@ -18,12 +28,13 @@ class MainActivity : ComponentActivity() {
         super.onCreate(savedInstanceState)
         setContent {
             PokemonDexTheme {
-                // A surface container using the 'background' color from the theme
                 Surface(
                     modifier = Modifier.fillMaxSize(),
                     color = MaterialTheme.colorScheme.background
                 ) {
-                    Greeting("Android")
+                    PokemonList(pokemons = PokemonData.pokemons, onPokemonClick = {
+                        // TODO: Implement this to navigate to the Pokemon details page
+                    })
                 }
             }
         }
@@ -31,17 +42,23 @@ class MainActivity : ComponentActivity() {
 }
 
 @Composable
-fun Greeting(name: String, modifier: Modifier = Modifier) {
-    Text(
-        text = "Hello $name!",
-        modifier = modifier
-    )
-}
-
-@Preview(showBackground = true)
-@Composable
-fun GreetingPreview() {
-    PokemonDexTheme {
-        Greeting("Android")
+fun PokemonList(pokemons: List<Pokemon>, onPokemonClick: (Pokemon) -> Unit) {
+    LazyColumn {
+        items(pokemons.size) { index ->
+            val pokemon = pokemons[index]
+            PokemonListItem(pokemon = pokemon, onClick = { onPokemonClick(pokemon) })
+        }
     }
 }
+
+@Composable
+fun PokemonListItem(pokemon: Pokemon, onClick: () -> Unit) {
+    Row(modifier = Modifier.clickable(onClick = onClick)) {
+        Image(painter = painterResource(pokemon.image), contentDescription = pokemon.name)
+        Column(modifier = Modifier.padding(start = 8.dp)) {
+            Text(text = pokemon.name)
+            Text(text = pokemon.id)
+        }
+    }
+}
+
